+++ autoswitch_virtualenv.plugin.zsh
@@ -1,13 +1,15 @@
-export AUTOSWITCH_VERSION="3.6.0"
+export AUTOSWITCH_VERSION="3.7.1"
 export AUTOSWITCH_FILE=".venv"
 export AUTOSWITCH_VIRTUAL_ENV_DIR="."
 export VIRTUAL_ENV=""
 
-RED="\e[31m"
-GREEN="\e[32m"
-PURPLE="\e[35m"
-BOLD="\e[1m"
-NORMAL="\e[0m"
+AUTOSWITCH_RED="\e[31m"
+AUTOSWITCH_GREEN="\e[32m"
+AUTOSWITCH_PURPLE="\e[35m"
+AUTOSWITCH_BOLD="\e[1m"
+AUTOSWITCH_NORMAL="\e[0m"
+
+VIRTUAL_ENV_DIR="${AUTOSWITCH_VIRTUAL_ENV_DIR:-$HOME/.virtualenvs}"
 
 function _validated_source() {
     local target_path="$1"
@@ -25,8 +27,6 @@ function _validated_source() {
 
 function _virtual_env_dir() {
     local venv_name="$1"
-    local VIRTUAL_ENV_DIR="${AUTOSWITCH_VIRTUAL_ENV_DIR:-$HOME/.virtualenvs}"
-    mkdir -p "$VIRTUAL_ENV_DIR"
     printf "%s/%s" "$VIRTUAL_ENV_DIR" "$venv_name"
 }
 
@@ -82,18 +82,18 @@ function _maybeworkon() {
     local venv_type="$2"
     local venv_name="$(_get_venv_name $venv_dir $venv_type)"
 
-    local DEFAULT_MESSAGE_FORMAT="Switching %venv_type: ${BOLD}${PURPLE}%venv_name${NORMAL} ${GREEN}[🐍%py_version]${NORMAL}"
+    local DEFAULT_MESSAGE_FORMAT="Switching %venv_type: ${AUTOSWITCH_BOLD}${AUTOSWITCH_PURPLE}%venv_name${AUTOSWITCH_NORMAL} ${AUTOSWITCH_GREEN}[🐍%py_version]${AUTOSWITCH_NORMAL}"
     if [[ "$LANG" != *".UTF-8" ]]; then
         # Remove multibyte characters if the terminal does not support utf-8
         DEFAULT_MESSAGE_FORMAT="${DEFAULT_MESSAGE_FORMAT/🐍/}"
     fi
 
     # Don't reactivate an already activated virtual environment
-    if [[ -z "$VIRTUAL_ENV" || "$venv_name" != "$(_get_venv_name $VIRTUAL_ENV $venv_type)" ]]; then
+    if [[ -z "$VIRTUAL_ENV" || "$venv_dir" != "$VIRTUAL_ENV" ]]; then
 
         if [[ ! -d "$venv_dir" ]]; then
-            printf "Unable to find ${PURPLE}$venv_name${NORMAL} virtualenv\n"
-            printf "If the issue persists run ${PURPLE}rmvenv && mkvenv${NORMAL} in this directory\n"
+            printf "Unable to find ${AUTOSWITCH_PURPLE}$venv_name${AUTOSWITCH_NORMAL} virtualenv\n"
+            printf "If the issue persists run ${AUTOSWITCH_PURPLE}rmvenv && mkvenv${AUTOSWITCH_NORMAL} in this directory\n"
             return
         fi
 
@@ -185,11 +185,11 @@ function check_venv()
         if [[ -f "$venv_path" ]] && [[ "$file_owner" != "$(id -u)" ]]; then
             printf "AUTOSWITCH WARNING: Virtualenv will not be activated\n\n"
             printf "Reason: Found a $AUTOSWITCH_FILE file but it is not owned by the current user\n"
-            printf "Change ownership of ${PURPLE}$venv_path${NORMAL} to ${PURPLE}'$USER'${NORMAL} to fix this\n"
+            printf "Change ownership of ${AUTOSWITCH_PURPLE}$venv_path${AUTOSWITCH_NORMAL} to ${PURPLE}'$USER'${NORMAL} to fix this\n"
         elif [[ -f "$venv_path" ]] && ! [[ "$file_permissions" =~ ^[64][04][04]$ ]]; then
             printf "AUTOSWITCH WARNING: Virtualenv will not be activated\n\n"
             printf "Reason: Found a $AUTOSWITCH_FILE file with weak permission settings ($file_permissions).\n"
-            printf "Run the following command to fix this: ${PURPLE}\"chmod 600 $venv_path\"${NORMAL}\n"
+            printf "Run the following command to fix this: ${AUTOSWITCH_PURPLE}\"chmod 600 $venv_path\"${AUTOSWITCH_NORMAL}\n"
         else
             if [[ "$venv_path" == *"/Pipfile" ]]; then
                 if type "pipenv" > /dev/null && _activate_pipenv; then
@@ -216,8 +216,8 @@ function check_venv()
 
     # If we still haven't got anywhere, fallback to defaults
     if [[ "$venv_type" != "unknown" ]]; then
-        printf "Python ${PURPLE}$venv_type${NORMAL} project detected. "
-        printf "Run ${PURPLE}mkvenv${NORMAL} to setup autoswitching\n"
+        printf "Python ${AUTOSWITCH_PURPLE}$venv_type${AUTOSWITCH_NORMAL} project detected. "
+        printf "Run ${AUTOSWITCH_PURPLE}mkvenv${AUTOSWITCH_NORMAL} to setup autoswitching\n"
     fi
     _default_venv
 }
@@ -231,7 +231,7 @@ function _default_venv()
         _maybeworkon "$(_virtual_env_dir "$AUTOSWITCH_DEFAULTENV")" "$venv_type"
     elif [[ -n "$VIRTUAL_ENV" ]]; then
         local venv_name="$(_get_venv_name "$VIRTUAL_ENV" "$venv_type")"
-        _autoswitch_message "Deactivating: ${BOLD}${PURPLE}%s${NORMAL}\n" "$venv_name"
+        _autoswitch_message "Deactivating: ${AUTOSWITCH_BOLD}${AUTOSWITCH_PURPLE}%s${AUTOSWITCH_NORMAL}\n" "$venv_name"
         deactivate
     fi
 }
@@ -260,12 +260,15 @@ function rmvenv()
                 fi
             fi
 
-            printf "Removing ${PURPLE}%s${NORMAL}...\n" "$venv_name"
+            printf "Removing ${AUTOSWITCH_PURPLE}%s${AUTOSWITCH_NORMAL}...\n" "$venv_name"
             # Using explicit paths to avoid any alias/function interference.
             # rm should always be found in this location according to
             # https://refspecs.linuxfoundation.org/FHS_3.0/fhs/ch03s04.html
             # https://www.freedesktop.org/wiki/Software/systemd/TheCaseForTheUsrMerge/
             /bin/rm -rf "$(_virtual_env_dir "$venv_name")"
+            if [[ -z "$(/bin/ls -A $VIRTUAL_ENV_DIR)" ]]; then
+                /bin/rm -rf "$VIRTUAL_ENV_DIR"
+            fi
             /bin/rm "$AUTOSWITCH_FILE"
         else
             printf "No $AUTOSWITCH_FILE file in the current directory!\n"
@@ -276,11 +279,11 @@ function rmvenv()
 
 function _missing_error_message() {
     local command="$1"
-    printf "${BOLD}${RED}"
+    printf "${AUTOSWITCH_BOLD}${AUTOSWITCH_RED}"
     printf "zsh-autoswitch-virtualenv requires '%s' to install this project!\n\n" "$command"
-    printf "${NORMAL}"
+    printf "${AUTOSWITCH_NORMAL}"
     printf "If this is already installed but you are still seeing this message, \n"
-    printf "then make sure the ${BOLD}$command${NORMAL} command is in your PATH.\n" $command
+    printf "then make sure the ${AUTOSWITCH_BOLD}$command${AUTOSWITCH_NORMAL} command is in your PATH.\n" $command
     printf "\n"
 }
 
@@ -330,16 +333,18 @@ function mkvenv()
             local venv_name="venv"
 
 
-            printf "Creating ${PURPLE}%s${NONE} virtualenv\n" "$venv_name"
+            printf "Creating ${AUTOSWITCH_PURPLE}%s${NONE} virtualenv\n" "$venv_name"
 
 
             if [[ -n "$AUTOSWITCH_DEFAULT_PYTHON" && ${params[(I)--python*]} -eq 0 ]]; then
-                printf "${PURPLE}"
+                printf "${AUTOSWITCH_PURPLE}"
                 printf 'Using $AUTOSWITCH_DEFAULT_PYTHON='
                 printf "$AUTOSWITCH_DEFAULT_PYTHON"
                 printf "${NONE}\n"
                 params+="--python=$AUTOSWITCH_DEFAULT_PYTHON"
             fi
+
+            /bin/mkdir -p "$VIRTUAL_ENV_DIR"
 
             if [[ ${params[(I)--verbose]} -eq 0 ]]; then
                 virtualenv $params "$(_virtual_env_dir "$venv_name")"
@@ -360,7 +365,7 @@ function mkvenv()
 
 function install_requirements() {
     if [[ -f "$AUTOSWITCH_DEFAULT_REQUIREMENTS" ]]; then
-        printf "Install default requirements? (${PURPLE}$AUTOSWITCH_DEFAULT_REQUIREMENTS${NORMAL}) [y/N]: "
+        printf "Install default requirements? (${AUTOSWITCH_PURPLE}$AUTOSWITCH_DEFAULT_REQUIREMENTS${AUTOSWITCH_NORMAL}) [y/N]: "
         read ans
 
         if [[ "$ans" = "y" || "$ans" == "Y" ]]; then
@@ -369,7 +374,7 @@ function install_requirements() {
     fi
 
     if [[ -f "$PWD/setup.py" ]]; then
-        printf "Found a ${PURPLE}setup.py${NORMAL} file. Install dependencies? [y/N]: "
+        printf "Found a ${AUTOSWITCH_PURPLE}setup.py${AUTOSWITCH_NORMAL} file. Install dependencies? [y/N]: "
         read ans
 
         if [[ "$ans" = "y" || "$ans" = "Y" ]]; then
@@ -386,7 +391,7 @@ function install_requirements() {
     local requirements
     for requirements in **/*requirements.txt
     do
-        printf "Found a ${PURPLE}%s${NORMAL} file. Install? [y/N]: " "$requirements"
+        printf "Found a ${AUTOSWITCH_PURPLE}%s${AUTOSWITCH_NORMAL} file. Install? [y/N]: " "$requirements"
         read ans
 
         if [[ "$ans" = "y" || "$ans" = "Y" ]]; then
@@ -412,15 +417,16 @@ function disable_autoswitch_virtualenv() {
 # immediately removes itself from the zsh-hook.
 # This seems important for "instant prompt" zsh themes like powerlevel10k
 function _autoswitch_startup() {
+    local python_bin="${AUTOSWITCH_DEFAULT_PYTHON:-python}"
+    if ! type "${python_bin}" > /dev/null; then
+        printf "WARNING: python binary '${python_bin}' not found on PATH.\n"
+        printf "zsh-autoswitch-virtualenv plugin will be disabled.\n"
+    else
+        enable_autoswitch_virtualenv
+        check_venv
+    fi
     add-zsh-hook -D precmd _autoswitch_startup
-    enable_autoswitch_virtualenv
-    check_venv
 }
 
-if ! type "${AUTOSWITCH_DEFAULT_PYTHON:-python}" > /dev/null; then
-    printf "WARNING: python binary not found on PATH.\n"
-    printf "zsh-autoswitch-virtualenv plugin will be disabled.\n"
-else
-    autoload -Uz add-zsh-hook
-    add-zsh-hook precmd _autoswitch_startup
-fi
+autoload -Uz add-zsh-hook
+add-zsh-hook precmd _autoswitch_startup